+++ Game/lib.h
@@ -3,7 +3,7 @@
 #define GRID_X 120
 #define GRID_Y 120
 
-int Lib_Rand(int min, int max);
+long Lib_Rand(int min, int max);
 void Lib_DrawCell(int x, int y, int color);
 void Lib_Display();
 int Lib_CountAliveNeighbors(int grid[GRID_Y][GRID_X], int x, int y);
